File Size 단위 수정 (Integer -> Double)
git-svn-id: https://skvalley.com/svn/repos/storeplatform/trunk/sac/storeplatform-sac@40031 01f5c84a-aead-4b29-b0f1-7c54dbeec05f

+++ src/main/java/com/skplanet/storeplatform/sac/display/epub/vo/EpubDetail.java
@@ -139,9 +139,9 @@ public class EpubDetail extends CommonInfo {
 
 	// Accrual
 	/** 참여자수 */
-	private Integer paticpersCnt;
+	private Double paticpersCnt;
 	/** 구매수 */
-	private Integer prchsCnt;
+	private Double prchsCnt;
 	/** 평균 평가 점수 */
 	private Double avgEvluScore;
